requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
openai==1.3.7
pinecone-client==2.2.4
pypdf2==3.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
networkx==3.2.1
d3py==0.0.1
matplotlib==3.8.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
